pyproject.toml
[tool.poetry]
name = "pipeline"
version = "0.1.0"
description = "This tool is part of the BChain4all project and aims to automate the process of creating and deploying smart contracts using LLM based on an external legal contract. The process is as automated as possible to ensure efficiency and accuracy."
authors = ["BChain4All"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12.0"
openai = "^1.10.0"
docker = "^7.0.0"
tiktoken = "^0.5.2"
mistralai = "^1.0.2"
nltk = "^3.8.1"
eth-brownie = {version = "^1.20.0", python = ">=3.10,<4.0"}
pandas = "^2.2.0"
google-generativeai = "^0.3.2"
slither-analyzer = "^0.10.3"
openpyxl = "^3.1.5"
anthropic = "^0.34.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
